Changelog: http://itextpdf.com/changelog/520
Release Notes: http://itextpdf.com/release/iText520
Download: https://github.com/itext/itextpdf/releases/tag/5.2.0

+++ itext/src/main/java/com/itextpdf/text/pdf/CFFFont.java
@@ -2,7 +2,7 @@
  * $Id$
  *
  * This file is part of the iText (R) project.
- * Copyright (c) 1998-2011 1T3XT BVBA
+ * Copyright (c) 1998-2012 1T3XT BVBA
  * Authors: Bruno Lowagie, Paulo Soares, et al.
  *
  * This program is free software; you can redistribute it and/or modify
@@ -248,7 +248,7 @@ public class CFFFont {
 
     int getPosition() {
         try {
-            return buf.getFilePointer();
+            return (int)buf.getFilePointer();
         }
         catch (Exception e) {
             throw new ExceptionConverter(e);
