+++ itext/src/main/java/com/itextpdf/text/pdf/PdfEFStream.java
@@ -2,7 +2,7 @@
  * $Id$
  *
  * This file is part of the iText (R) project.
- * Copyright (c) 1998-2011 1T3XT BVBA
+ * Copyright (c) 1998-2012 1T3XT BVBA
  * Authors: Bruno Lowagie, Paulo Soares, et al.
  *
  * This program is free software; you can redistribute it and/or modify
@@ -148,7 +148,7 @@ public class PdfEFStream extends PdfStream {
             }
             if (ose != null)
                 ose.finish();
-            inputStreamLength = osc.getCounter();
+            inputStreamLength = (int)osc.getCounter();
         }
         else {
             if (crypto == null) {
