+++ itext/src/main/java/com/itextpdf/text/pdf/Type1Font.java
@@ -2,7 +2,7 @@
  * $Id$
  *
  * This file is part of the iText (R) project.
- * Copyright (c) 1998-2011 1T3XT BVBA
+ * Copyright (c) 1998-2012 1T3XT BVBA
  * Authors: Bruno Lowagie, Paulo Soares, et al.
  *
  * This program is free software; you can redistribute it and/or modify
@@ -507,7 +507,7 @@ class Type1Font extends BaseFont
                 rf = new RandomAccessFileOrArray(filePfb, true, Document.plainRandomAccess);
             else
                 rf = new RandomAccessFileOrArray(pfb);
-            int fileLength = rf.length();
+            int fileLength = (int)rf.length();
             byte st[] = new byte[fileLength - 18];
             int lengths[] = new int[3];
             int bytePtr = 0;
